requirements.txt
flask==2.0.1
openai==1.3.0
python-dotenv==0.19.1
langchain==0.1.4
langchain-openai==0.0.5
langchain-community==0.0.13
faiss-cpu==1.7.4
PyPDF2==3.0.1
twilio==8.2.0
python-docx==0.8.11
elevenlabs==0.2.24
gunicorn==20.1.0
boto3==1.34.11
werkzeug==2.0.1